limited -> counted
