+++ topicsdb/search_test.go
@@ -5,19 +5,15 @@ import (
 	"testing"
 
 	"github.com/Fantom-foundation/lachesis-base/inter/idx"
-	"github.com/Fantom-foundation/lachesis-base/kvdb/memorydb"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/stretchr/testify/require"
-
-	"github.com/Fantom-foundation/go-opera/kvdb/threads"
 )
 
 func BenchmarkSearch(b *testing.B) {
 	topics, recs, topics4rec := genTestData(1000)
 
-	mem := threads.LimitedDBProducer(memorydb.NewProducer(""))
-	index := newIndex(mem)
+	index := newTestIndex()
 
 	for _, rec := range recs {
 		err := index.Push(rec)
